Improvements in auto_argmatcher:
Carapace is a tool that can generate completions for over 650 different
commands.  But some of the commands it recognizes are Unix or Linux
commands which can conflict with Windows commands by the same name.

For carapace to work, it has to hook into Clink to recognize command
names even before a space is pressed after the command name.  That
prevented auto_argmatcher from being able to override completions for
Windows commands that carapace gets wrong.

With the change here, auto_argmatcher has a chance to load an argmatcher
for a command earlier than carapace, **IF** the auto_argmatcher.config
file uses a fully qualified path for the command.  For example, I have a
custom rm.exe program I wrote, and I override carapace for it like this
in the auto_argmatcher.config file:

    c:\my_tools\rm.exe    -?    --slashes --concat

FOR MORE INFO ON USING CARAPACE WITH CLINK, REFER TO CLINK ISSUE 801.